Rename footer sections to increase clarity
