+++ src/components/Resources.svelte
