+++ src/components/TrustedBy.svelte
@@ -1,5 +1,5 @@
 <style>
-  .trusted {
+  .trustedby {
     background-color: #e7e7e7;
   }
 
@@ -7,37 +7,37 @@
     max-width: 1200px;
   }
 
-  .trusted__title-wrapper {
+  .trustedby__title-wrapper {
     display: flex;
     align-items: center;
   }
 
-  .trusted__title {
+  .trustedby__title {
     margin: 0;
     font-size: 18px;
     font-weight: bold;
     color: #393939;
   }
 
-  .trusted__img-wrapper {
+  .trustedby__img-wrapper {
     display: flex;
     align-items: center;
     height: 100%;
     max-width: 200px;
   }
 
-  .trusted__img {
+  .trustedby__img {
     max-width: 100%;
     max-height: 100px;
   }
 
   @media screen and (min-width: 768px) {
-    .trusted {
+    .trustedby {
       padding-top: 80px;
       padding-bottom: 85px;
     }
 
-    .trusted__title-wrapper {
+    .trustedby__title-wrapper {
       flex: 0 0 20%;
       max-width: 20%;
     }
@@ -49,7 +49,7 @@
   }
 
   @media screen and (max-width: 768px) {
-    .trusted {
+    .trustedby {
       padding-top: 25px;
       padding-bottom: 80px;
     }
@@ -59,11 +59,11 @@
       padding-left: 18px;
     }
 
-    .trusted__title-wrapper {
+    .trustedby__title-wrapper {
       justify-content: center;
     }
 
-    .trusted__img-wrapper {
+    .trustedby__img-wrapper {
       padding-top: 67px;
       margin: 0 auto;
       justify-content: center;
@@ -71,33 +71,33 @@
   }
 </style>
 
-<section class="trusted">
+<section class="trustedby">
   <div class="container-fluid">
     <div class="row">
-      <div class="trusted__title-wrapper col-12">
-        <h3 class="trusted__title">Trusted by</h3>
+      <div class="trustedby__title-wrapper col-12">
+        <h3 class="trustedby__title">Trusted by</h3>
       </div>
       <div class="col-12 col-md-3">
-        <div class="trusted__img-wrapper">
+        <div class="trustedby__img-wrapper">
           <img
-            class="trusted__img"
-            src="trusted/hkpc-logo.png"
+            class="trustedby__img"
+            src="trustedby/hkpc-logo.png"
             alt="HKPC Logo" />
         </div>
       </div>
       <div class="col-12 col-md-3">
-        <div class="trusted__img-wrapper">
+        <div class="trustedby__img-wrapper">
           <img
-            class="trusted__img"
-            src="trusted/cmic-logo.png"
+            class="trustedby__img"
+            src="trustedby/cmic-logo.png"
             alt="CMIC Logo" />
         </div>
       </div>
       <div class="col-12 col-md-3">
-        <div class="trusted__img-wrapper">
+        <div class="trustedby__img-wrapper">
           <img
-            class="trusted__img"
-            src="trusted/rees-logo.png"
+            class="trustedby__img"
+            src="trustedby/rees-logo.png"
             alt="REES Logo" />
         </div>
       </div>
